requirements.txt
# Core Dependencies
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Async & Networking
asyncio>=3.4.3
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.25.0

# Data & APIs
yfinance>=0.2.33
finnhub-python>=2.4.18
vaderSentiment>=3.3.2

# Visualization
plotly>=5.18.0

# Utilities
pydantic>=2.5.0
click>=8.1.0
colorama>=0.4.6
tenacity>=8.2.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0